Add ControlNet task tests and improve temp path cleanup

+++ tests/back/tasks/test_tasks.py
@@ -12,6 +12,7 @@ from DashAI.back.dataloaders.classes.dashai_dataset import (
     to_dashai_dataset,
 )
 from DashAI.back.dataloaders.classes.json_dataloader import JSONDataLoader
+from DashAI.back.tasks.controlnet_task import ControlNetTask
 from DashAI.back.tasks.tabular_classification_task import TabularClassificationTask
 from DashAI.back.tasks.text_classification_task import TextClassificationTask
 from DashAI.back.tasks.text_to_image_generation_task import TextToImageGenerationTask
@@ -251,9 +252,13 @@ def temp_path_fixture():
     os.makedirs(temp_path, exist_ok=True)
     yield temp_path
     # Cleanup after all tests in the module using this fixture have finished
-    for file in os.listdir(temp_path):
-        os.remove(os.path.join(temp_path, file))
-    os.rmdir(temp_path)
+    if temp_path.exists() and temp_path.is_dir():
+        for root, dirs, files in os.walk(temp_path, topdown=False):
+            for file in files:
+                os.remove(os.path.join(root, file))
+            for dir in dirs:
+                os.rmdir(os.path.join(root, dir))
+        os.rmdir(temp_path)
 
 
 # Text To Text Generation Task
@@ -372,3 +377,72 @@ def test_process_input_from_database_text_to_image():
     processed_input = text_to_image_task.process_input_from_database(input_data)
 
     assert processed_input == input_data
+
+
+# ControlNet Task
+def test_get_controlnet_task_metadata():
+    controlnet_task = ControlNetTask()
+    metadata = controlnet_task.get_metadata()
+
+    assert len(metadata.keys()) == 4
+    assert metadata["inputs_types"] == ["Image", "str"]
+    assert metadata["outputs_types"] == ["Image"]
+    assert metadata["inputs_cardinality"] == 2
+    assert metadata["outputs_cardinality"] == "n"
+
+
+def test_prepare_for_task_controlnet(sample_image, temp_path):
+    controlnet_task = ControlNetTask()
+    path_image = temp_path / "sample_controlnet_image.png"
+    sample_image.save(path_image)
+
+    input_data = (path_image, "A beautiful landscape")
+
+    prepared_input = controlnet_task.prepare_for_task(input_data)
+
+    assert isinstance(prepared_input, list)
+    assert isinstance(prepared_input[0], PIL.Image.Image)
+    assert isinstance(prepared_input[1], str)
+
+
+def test_prepare_input_for_database_controlnet(sample_image, temp_path):
+    controlnet_task = ControlNetTask()
+
+    bytes_image = sample_image.tobytes()
+
+    input_data = (bytes_image, "A beautiful landscape")
+    prepared_input = controlnet_task.prepare_input_for_database(
+        input_data, images_path=temp_path / "controlnet_prepare_for_database"
+    )
+
+    assert isinstance(prepared_input, list)
+    assert isinstance(prepared_input[0], str)
+    assert isinstance(prepared_input[1], str)
+    assert prepared_input[1] == "A beautiful landscape"
+    # Check if the image is saved in the temp path
+    assert len(os.listdir(temp_path / "controlnet_prepare_for_database")) == 1
+
+
+def test_process_output_controlnet(sample_image, temp_path):
+    controlnet_task = ControlNetTask()
+    output_data = [sample_image]
+    processed_output = controlnet_task.process_output(
+        output_data, images_path=temp_path / "controlnet_process_output"
+    )
+
+    assert isinstance(processed_output, list)
+    assert all(isinstance(img, str) for img in processed_output)
+    # Check if the image is saved in the temp path
+    assert len(os.listdir(temp_path / "controlnet_process_output")) == 1
+
+
+def test_process_output_from_database_controlnet():
+    controlnet_task = ControlNetTask()
+    output_data = ["dir/sample_image_0.png", "dir/sample_image_1.png"]
+    processed_output = controlnet_task.process_output_from_database(output_data)
+
+    assert isinstance(processed_output, list)
+    assert all(isinstance(img, str) for img in processed_output)
+    assert len(processed_output) == 2
+    assert processed_output[0] == "sample_image_0.png"
+    assert processed_output[1] == "sample_image_1.png"
